% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCFunctions.R
\name{ComputeROCFunctions}
\alias{ComputeROCFunctions}
\title{Compute ROC functions}
\usage{
ComputeROCFunctions(x, rocs = c("FPR", "TPR"))
}
\arguments{
\item{x}{A numeric vector or data frame with names:
\code{ c("TruePositive", "TrueNegative", "FalsePositive", "FalseNegative" )}.}

\item{rocs}{A list of ROC functions names}
}
\value{
A data frame
}
\description{
Computes specified Receiver Operating Characteristic (ROC)
functions over a list or a data frame.
}
