% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCFunctions.R
\name{ComputeROCTypeCounts}
\alias{ComputeROCTypeCounts}
\title{Compute ROC type counts}
\usage{
ComputeROCTypeCounts(
  data,
  actualColumnName,
  probColumnName,
  thresholds = seq(0, 1, 0.05),
  falseLabel = "0",
  trueLabel = "1",
  form = "wide"
)
}
\arguments{
\item{data}{A data frame}

\item{actualColumnName}{A string with the actual labels column name.}

\item{probColumnName}{A string with the column name of the probabilities
to assign the "true" label.}

\item{thresholds}{A numeric vector with numbers between 0 and 1.}

\item{falseLabel}{A string that is "false" label.}

\item{trueLabel}{A string that is "true" label.}
}
\value{
A data frame
}
\description{
Computes core Receiver Operating Characteristic (ROC)
counts over a data frame with predictions data.
}
\details{
If \code{form == "wide"}
them it produces a data frame with the columns
\code{ c("Threshold", "TruePositive", "TrueNegative", "FalsePositive", "FalseNegative" )}.
Otherwise the long form version is returned.
}
