% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCFunctions.R
\name{ROCAcronymsDictionary}
\alias{ROCAcronymsDictionary}
\title{ROC acronyms dictionary}
\usage{
ROCAcronymsDictionary()
}
\value{
A data frame
}
\description{
Gives a data frame that is a ROC acronyms dictionary.
}
