% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCFunctions.R
\name{ROCDataFrameQ}
\alias{ROCDataFrameQ}
\title{ROCDataFrameQ}
\usage{
ROCDataFrameQ(x)
}
\arguments{
\item{x}{Object}
}
\value{
A logical value
}
\description{
Is an object a data frame with ROC columns or not.
}
